update test sdk plugins
